refactor: remove unused files from utils
